saidas ordenas no menor caminho para o maior, e testador e comparador de saidas

+++ src/mainMatrix.c
@@ -20,7 +20,7 @@ int main(int argc, char *argv[])
     resolveComDijkstraMatrix(entradaPath, saidaPath);
 
     gettimeofday(&end, NULL); // Marca o tempo final
-    double elapsed_time = (end.tv_sec - start.tv_sec) + (end.tv_usec - start.tv_usec) / 1000000.0;
+    float elapsed_time = (end.tv_sec - start.tv_sec) + (end.tv_usec - start.tv_usec) / 1000000.0;
     printf("Tempo de execução (matrix) [%s]: %.6f segundos\n", entradaPath, elapsed_time);
 
     return 0;
